requirements.txt
asgiref==3.11.0
boto3==1.42.25
botocore==1.42.25
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.10
django-cors-headers==4.9.0
django-filter==25.2
django-js-asset==3.1.2
django-mptt==0.18.0
django-storages==1.14.6
django-utils==0.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
jmespath==1.0.1
mysqlclient==2.2.7
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
pycparser==2.23
PyJWT==2.10.1
PyMuPDF==1.26.7
pyotp==2.9.0
pypdf==6.6.2
pypdfium2==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
qrcode==8.2
reportlab==4.4.9
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
